+++ js/components/PluginList.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { connect } from 'react-redux';
 import PluginListItem from './PluginListItem';
-import VisuallyHidden from './VisuallyHidden.react';
+import VisuallyHidden from './VisuallyHidden';
 import Spinner from './Spinner';
 
 const PluginList = ({ plugins, location, route }) => {
